requirements.txt
absl-py==2.1.0
anyio==4.4.0
archspec==0.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
boltons==23.0.0
Bottleneck==1.4.2
Brotli==1.0.9
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==2.0.4
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
einops==0.8.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.53.0
fqdn==1.5.1
frozendict==2.4.2
fsspec==2024.5.0
gmpy2==2.2.1
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.30.2
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab-language-pack-zh-CN==4.2.post1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
ml_collections==1.1.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook_shim==0.2.4
numexpr==2.10.1
numpy==2.2.1
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pip==24.0
platformdirs==3.10.0
pluggy==1.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.45
protobuf==5.27.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycosat==0.6.6
pycparser==2.21
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
rpds-py==0.18.1
ruamel.yaml==0.17.21
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
Send2Trash==1.8.3
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
supervisor==4.2.5
sympy==1.13.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.5.0
timm==1.0.15
tinycss2==1.3.0
tokenizers==0.21.1
torch-pruning==1.5.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.51.3
truststore==0.8.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
zstandard==0.22.0
